support rails 5 attributes api
When we call `obj.attributes` on an ActiveRecord object, it will return all of the model’s attributes, including those defined with the Rails 5 attributes API. 

Here, `fixture_builder` will build fixture files and include attributes that are not backed by columns. This causes issues when the fixtures are loaded by RSpec.

Now, we only consider column names when building fixtures.